break run in several methods, execute user_cmd abort when user is in state connected

+++ ext/QVD-HKD/lib/QVD/HKD.pm
@@ -18,197 +18,214 @@ use POSIX;
 use Log::Log4perl qw(:levels :easy);
 Log::Log4perl::init('log4perl.conf');
 
-my %timeout = ( starting   => cfg(vm_state_starting_timeout,   200),
-	        stopping_1 => cfg(vm_state_stopping_1_timeout,  30),
-		stopping_2 => cfg(vm_state_stopping_2_timeout, 200),
-		zombie_1   => cfg(vm_state_zombie_1_timeout,    30) );
+my %timeout = ( starting    => cfg(vm_state_starting_timeout,    200),
+	        stopping_1  => cfg(vm_state_stopping_1_timeout,   30),
+		stopping_2  => cfg(vm_state_stopping_2_timeout,  200),
+		zombie_1    => cfg(vm_state_zombie_1_timeout,     30),
+		vma         => cfg(vm_state_running_vma_timeout, 120) );
 
 # FIXME: read nodename from configuration file!
 my $this_host_id = rs(Host)->search(name => hostname)->first->id;
 
-sub _reap_children { 1 while (waitpid(-1, WNOHANG) > 0) }
-
-sub new {
-    my ($class, %opts) = @_;
-    my $self = {};
-    bless $self, $class;
-    $self;
-}
+# The class QVD::HKD does not have state so we use the class name as
+# the object.
+#
+# sub new { ... }
 
 sub run {
-    my $self = shift;
+    my $hkd = shift;
 
-    my $vma_ok_timeout = cfg(vm_state_running_vma_timeout);
-
+    # flag to test VMAs on running machines once every twelve runs
     my $round = 0;
 
     while (1) {
-        # FIXME: make this block into another method and wrap it
-        # inside an eval
-
 	DEBUG "HKD run, round: $round";
 
-	$self->_reap_children;
+	$hkd->_reap_children;
 
-	my $par = QVD::ParallelNet->new;
+	my $check_all = not $round++ % 12;
+	$hkd->_check_vms($check_all);
 
-	my @vms = rs(VM_Runtime)->search({host_id => $this_host_id});
-	my (@active_vms, @vmas);
+	$hkd->_check_l7rs;
 
-	for my $vm (@vms) {
-	    my $id = $vm->id;
-
-	    # Command processing...
-	    my $start;
-	    if (defined $vm->vm_cmd) {
-		txn_eval {
-		    $vm->discard_changes;
-		    given($vm->vm_cmd) {
-			when('start') {
-			    given($vm->vm_state) {
-				when ('stopped') {
-				    $self->_assign_vm_ports($vm);
-				    $self->_move_vm_to_state(starting => $vm);
-                                    $vm->clear_vm_cmd;
-				    $start = 1;
-				}
-				default {
-				    ERROR "unexpected VM command start received in state $_";
-				    $vm->clear_vm_cmd;
-				}
-			    }
-			}
-			when('stop') {
-			    given($vm->vm_state) {
-				when ('running')  {
-                                    $self->_move_vm_to_state(stopping_1 => $vm);
-                                    $vm->clear_vm_cmd;
-                                }
-				when ('starting') { } # stop is delayed!
-				default {
-				    ERROR "unexpected VM command stop received in state $_";
-				    $vm->clear_vm_cmd;
-				}
-			    }
-			}
-			when(undef) {
-			    DEBUG "command dissapeared";
-			}
-			default {
-			    ERROR "unexpected VM command $_ received in state " . $vm->vm_state;
-                            $vm->clear_vm_cmd;
-			}
-		    }
-		};
-		$@ and ERROR "vm_cmd processing failed: $@";
-	    }
-
-	    # no error checking is performed here, failed virtual
-	    # machines startings are captured later or on the next
-	    # run:
-	    if ($start) {
-		eval { $self->_start_vm($vm) };
-		$@ and ERROR "Unable to start VM: $@";
-	    }
-
-	    next if $vm->vm_state eq 'stopped';
-
-	    unless ($self->_check_vm_process($vm)) {
-                given ($vm->vm_state) {
-                    when ('stopping_1') {}
-                    when ('stopping_2') {
-                        WARN "vm process exited without passing through stopping_2"
-                    }
-                    default {
-                        ERROR "vm process has disappeared!, id: $id";
-                        $vm->block;
-                    }
-		}
-		txn_eval { $self->_move_vm_to_state(stopped => $vm) };
-		$@ and ERROR "unable to move VM $id to state stopped";
-
-		next;
-	    }
-
-	    my $vma_method;
-	    given ($vm->vm_state) {
-		when('running'   ) { $vma_method = 'status' unless $round }
-		when('starting'  ) { $vma_method = 'status' }
-		when('stopping_1') { $vma_method = 'poweroff' }
-		when('zombie_1'  ) { $self->_signal_vm($vm, SIGTERM) }
-		when('zombie_2'  ) { $self->_signal_vm($vm, SIGKILL) }
-	    }
-
-	    if (defined $vma_method) {
-		my $vma = QVD::SimpleRPC::Client::Parallel->new($vm->vma_url);
-		$vma->queue_request($vma_method);
-		$par->register($vma);
-		push @vmas, $vma;
-		push @active_vms, $vm;
-	    }
-	    else {
-		$self->_go_zombie_on_timeout($vm);
-	    }
-	}
-
-	$par->run(time => 2) if @active_vms;
-
-	while (@active_vms) {
-	    my $vm = shift @active_vms;
-	    my $vma = shift @vmas;
-	    my $id = $vm->id;
-	    eval { $vma->unqueue_response };
-	    if ($@) {
-		DEBUG "VMA call in VM $id failed: $@";
-		given ($vm->vm_state) {
-		    when('running') {
-			my $vma_ok_ts = $vm->vma_ok_ts;
-			DEBUG "vma_timeout $vma_ok_timeout, elapsed " . (time - $vma_ok_ts);
-			if (max($^T, $vma_ok_ts) + $vma_ok_timeout < time) {
-			    # FIXME: check also that the number of consecutive
-			    # failed checks goes over some threshold
-			    ERROR "machine has not responded for a long time (" .
-				(time - $vma_ok_ts) . " seconds), going zombie!" .
-				    " id: $id, vma_ok_ts: $vma_ok_ts, time: " . time;
-			    txn_eval { $self->_move_vm_to_state(zombie_1 => $vm) };
-			    $@ and ERROR "unable to move VM to state zombie_1";
-			}
-		    }
-		    default {
-			$self->_go_zombie_on_timeout($vm);
-		    }
-		}
-	    }
-	    else {
-		$vm->update_vma_ok_ts;
-
-		my $new_state;
-		given ($vm->vm_state) {
-		    when ('starting')   { $new_state = 'running' }
-		    when ('stopping_1') { $new_state = 'stopping_2' }
-		}
-		if (defined $new_state) {
-		    txn_eval { $self->_move_vm_to_state($new_state => $vm) };
-		    $@ and ERROR "Unable to move VM $id to state $new_state: $@";
-		}
-	    }
-	}
-
-	# check for dissapearing L7Rs processes
-	my @l7rs = rs(VM_Runtime)->search({l7r_host => $this_host_id});
-	for (@l7rs) {
-            $_->clear_l7r_all
-                unless $self->_check_l7r_process($_);
-        }
-
-	# flag to test VMAs on running machines once every twelve runs
-	$round = (($round + 1) % 12);
 	sleep 2;
     }
 }
 
+sub _reap_children { 1 while (waitpid(-1, WNOHANG) > 0) }
+
+sub _check_vms {
+    my ($hkd, $check_all) = @_;
+
+    my (@active_vms, @vmas);
+    my $par = QVD::ParallelNet->new;
+
+    for my $vm (rs(VM_Runtime)->search({host_id => $this_host_id})) {
+	my $id = $vm->id;
+
+	# Command processing...
+	my $start;
+	if (defined $vm->vm_cmd) {
+	    txn_eval {
+		$vm->discard_changes;
+		given($vm->vm_cmd) {
+		    when('start') {
+			given($vm->vm_state) {
+			    when ('stopped') {
+				$hkd->_assign_vm_ports($vm);
+				$hkd->_move_vm_to_state(starting => $vm);
+				$vm->clear_vm_cmd;
+				$start = 1;
+			    }
+			    default {
+				ERROR "unexpected VM command start received in state $_";
+				$vm->clear_vm_cmd;
+			    }
+			}
+		    }
+		    when('stop') {
+			given($vm->vm_state) {
+			    when ('running')  {
+				$hkd->_move_vm_to_state(stopping_1 => $vm);
+				$vm->clear_vm_cmd;
+			    }
+			    when ('starting') { } # stop is delayed!
+			    default {
+				ERROR "unexpected VM command stop received in state $_";
+				$vm->clear_vm_cmd;
+			    }
+			}
+		    }
+		    when(undef) {
+			DEBUG "command dissapeared";
+		    }
+		    default {
+			ERROR "unexpected VM command $_ received in state " . $vm->vm_state;
+			$vm->clear_vm_cmd;
+		    }
+		}
+	    };
+	    $@ and ERROR "vm_cmd processing failed: $@";
+	}
+
+	# no error checking is performed here, failed virtual
+	# machines startings are captured later or on the next
+	# run:
+	if ($start) {
+	    eval { $hkd->_start_vm($vm) };
+	    $@ and ERROR "Unable to start VM: $@";
+	}
+
+	next if $vm->vm_state eq 'stopped';
+
+	unless ($hkd->_check_vm_process($vm)) {
+	    given ($vm->vm_state) {
+		when ('stopping_1') {}
+		when ('stopping_2') {
+		    WARN "vm process exited without passing through stopping_2"
+		}
+		default {
+		    ERROR "vm process has disappeared!, id: $id";
+		    $vm->block;
+		}
+	    }
+	    txn_eval { $hkd->_move_vm_to_state(stopped => $vm) };
+	    $@ and ERROR "unable to move VM $id to state stopped";
+	    next;
+	}
+
+	my $vma_method;
+	given ($vm->vm_state) {
+	    when('running') {
+		no warnings 'uninitialized';
+		if ($vm->user_cmd eq 'abort' and $vm->user_state eq 'connected') {
+		    # HKD does this on behalf of the contending L7R to
+		    # handle the case where the abort is sent using
+		    # the administration tools.
+		    #
+		    # There is a race condition here: if the former
+		    # connection closes and the new is stablished
+		    # before the message is delivered the new one will
+		    # be aborted. This is very unlikely and mostly
+		    # harmless, so we didn't care!
+		    $vma_method = 'disconnect_session';
+		}
+		elsif ($check_all) {
+		    $vma_method = 'status';
+		}
+	    }
+	    when('starting'  ) { $vma_method = 'status' }
+	    when('stopping_1') { $vma_method = 'poweroff' }
+	    when('zombie_1'  ) { $hkd->_signal_vm($vm, SIGTERM) }
+	    when('zombie_2'  ) { $hkd->_signal_vm($vm, SIGKILL) }
+	}
+
+	if (defined $vma_method) {
+	    my $vma = QVD::SimpleRPC::Client::Parallel->new($vm->vma_url);
+	    $vma->queue_request($vma_method);
+	    $par->register($vma);
+	    push @vmas, $vma;
+	    push @active_vms, $vm;
+	}
+	else {
+	    $hkd->_go_zombie_on_timeout($vm);
+	}
+    }
+
+    $par->run(time => 2) if @active_vms;
+
+    while (@active_vms) {
+	my $vm = shift @active_vms;
+	my $vma = shift @vmas;
+	my $id = $vm->id;
+	eval { $vma->unqueue_response };
+	if ($@) {
+	    DEBUG "VMA call in VM $id failed: $@";
+	    given ($vm->vm_state) {
+		when('running') {
+		    my $vma_ok_ts = $vm->vma_ok_ts;
+		    DEBUG "vma_timeout $timeout{vma}, elapsed " . (time - $vma_ok_ts);
+		    if (max($^T, $vma_ok_ts) + $timeout{vma} < time) {
+			# FIXME: check also that the number of consecutive
+			# failed checks goes over some threshold
+			ERROR "machine has not responded for a long time (" .
+			    (time - $vma_ok_ts) . " seconds), going zombie!" .
+				" id: $id, vma_ok_ts: $vma_ok_ts, time: " . time;
+			txn_eval { $hkd->_move_vm_to_state(zombie_1 => $vm) };
+			$@ and ERROR "unable to move VM to state zombie_1";
+		    }
+		}
+		default {
+		    $hkd->_go_zombie_on_timeout($vm);
+		}
+	    }
+	}
+	else {
+	    $vm->update_vma_ok_ts;
+	    my $new_state;
+	    given ($vm->vm_state) {
+		when ('starting')   { $new_state = 'running' }
+		when ('stopping_1') { $new_state = 'stopping_2' }
+	    }
+	    if (defined $new_state) {
+		txn_eval { $hkd->_move_vm_to_state($new_state => $vm) };
+		$@ and ERROR "Unable to move VM $id to state $new_state: $@";
+	    }
+	}
+    }
+}
+
+sub _check_l7rs {
+    my $hkd = shift;
+    # check for dissapearing L7Rs processes
+    for my $l7r (rs(VM_Runtime)->search({l7r_host => $this_host_id})) {
+	$l7r->clear_l7r_all
+	    unless $hkd->_check_l7r_process($l7r);
+    }
+}
+
 sub _go_zombie_on_timeout {
-    my ($self, $vm) = @_;
+    my ($hkd, $vm) = @_;
     my $vm_state = $vm->vm_state;
     my $id = $vm->id;
     my $timeout = $timeout{$vm_state};
@@ -220,7 +237,7 @@ sub _go_zombie_on_timeout {
 	    ERROR "vm staled in state $vm_state,".
 		" id: $id, state_ts: $vm_state_ts, time: ".time;
 	    my $new_state = ($vm_state eq 'zombie_1' ? 'zombie_2' : 'zombie_1');
-	    eval { $self->_move_vm_to_state($new_state => $vm) };
+	    eval { $hkd->_move_vm_to_state($new_state => $vm) };
 	    $@ and ERROR "Unable to move VM $id to state $new_state: $@";
 	}
     }
@@ -232,53 +249,53 @@ my $port = 2000;
 sub _allocate_port { $port++ }
 
 sub _assign_vm_ports {
-    my ($self, $vm) = @_;
+    my ($hkd, $vm) = @_;
     $vm->update({vm_address => $vm->host->address,
-		 vm_vma_port => $self->_allocate_port,
-		 vm_x_port => $self->_allocate_port,
-		 vm_ssh_port => $self->_allocate_port,
+		 vm_vma_port => $hkd->_allocate_port,
+		 vm_x_port => $hkd->_allocate_port,
+		 vm_ssh_port => $hkd->_allocate_port,
 		 vm_vnc_port => 5900 + $vm->vm_id });
 }
 
 # this method must always be called from inside a txn_eval block!!!
 sub _move_vm_to_state {
-    my ($self, $vm_state, $vm) = @_;
+    my ($hkd, $vm_state, $vm) = @_;
     my $old_vm_state = $vm->vm_state;
     my $id = $vm->id;
     DEBUG "move VM $id from state $old_vm_state to $vm_state";
 
-    my $leave = $self->can("_leave_vm_state_$old_vm_state")
-	or DEBUG "method _leave_vm_state_$old_vm_state does not exist";
-    my $enter = $self->can("_enter_vm_state_$vm_state")
-	or DEBUG "method _enter_vm_state_$vm_state does not exist";
+    my $leave = $hkd->can("_leave_vm_state_$old_vm_state");
+    # or DEBUG "method _leave_vm_state_$old_vm_state does not exist";
+    my $enter = $hkd->can("_enter_vm_state_$vm_state");
+    # or DEBUG "method _enter_vm_state_$vm_state does not exist";
 
-    $leave->($self, $vm) if $leave;
-    $enter->($self, $vm) if $enter;
+    $leave->($hkd, $vm) if $leave;
+    $enter->($hkd, $vm) if $enter;
     $vm->set_vm_state($vm_state);
 }
 
 sub _leave_vm_state_running {
-    my ($self, $vm) = @_;
+    my ($hkd, $vm) = @_;
     $vm->clear_vma_ok_ts;
 }
 
 sub _enter_vm_state_stopping_1 {
-    my ($self, $vm) = @_;
+    my ($hkd, $vm) = @_;
     $vm->send_user_cmd('abort') if $vm->user_state eq 'connecting';
 }
 
 sub _enter_vm_state_stopped {
-    my ($self, $vm) = @_;
+    my ($hkd, $vm) = @_;
     $vm->clear_host_id;
 }
 
 sub _enter_vm_state_zombie_1 {
-    my ($self, $vm) = @_;
+    my ($hkd, $vm) = @_;
     $vm->send_user_cmd('abort') if $vm->user_state eq 'connecting';
 }
 
 sub _start_vm {
-    my ($self, $vm) = @_;
+    my ($hkd, $vm) = @_;
     my $id = $vm->vm_id;
     my $vma_port = $vm->vm_vma_port;
     my $x_port = $vm->vm_x_port;
@@ -295,14 +312,14 @@ sub _start_vm {
                -redir => "tcp:${vma_port}::3030",
                -redir => "tcp:${ssh_port}::22");
 
-    my $image = $self->_vm_image_path($vm, 1) //
+    my $image = $hkd->_vm_image_path($vm, 1) //
 	die "no disk image for vm $id";
 
     DEBUG "Using image $image for VM $id";
     push @cmd, -hda => $image;
 
     if (defined $osi->user_storage_size) {
-        my $user_storage = $self->_vm_user_storage_path($vm, 1) //
+        my $user_storage = $hkd->_vm_user_storage_path($vm, 1) //
             die "no user storage for vm $id";
 
 	DEBUG "Using user storage $user_storage for VM $id";
@@ -321,7 +338,7 @@ sub _start_vm {
 }
 
 sub _vm_image_path {
-    my ($self, $vm, $create_if_needed) = @_;
+    my ($hkd, $vm, $create_if_needed) = @_;
     my $id = $vm->id;
     my $osi = $vm->rel_vm_id->osi;
     my $osiid = $osi->id;
@@ -358,7 +375,7 @@ sub _vm_image_path {
 }
 
 sub _vm_user_storage_path {
-    my ($self, $vm, $create_if_needed) = @_;
+    my ($hkd, $vm, $create_if_needed) = @_;
     my $id = $vm->id;
     my $osi = $vm->rel_vm_id->osi;
     my $size = $osi->user_storage_size // return undef;
@@ -383,21 +400,21 @@ sub _vm_user_storage_path {
 }
 
 sub _signal_vm {
-    my ($self, $vm, $signal) = @_;
+    my ($hkd, $vm, $signal) = @_;
     my $pid = $vm->vm_pid;
-    DEBUG "kill VM process $pid with signal $signal";
+    DEBUG "kill VM process $pid with signal $signal" if $signal;
     kill($signal, $pid);
 }
 
 sub _check_vm_process {
-    my ($self, $vm) = @_;
-    $self->_signal_vm($vm, 0);
+    my ($hkd, $vm) = @_;
+    $hkd->_signal_vm($vm, 0);
 }
 
 sub _check_l7r_process {
-    my ($self, $vm) = @_;
+    my ($hkd, $vm) = @_;
     my $pid = $vm->l7r_pid;
-    DEBUG "kill L7R process $pid with signal 0";
+    # DEBUG "kill L7R process $pid with signal 0";
     kill(0, $pid);
 }
